qr_server/requirements.txt
Flask==2.3.3
opencv-python
qreader==3.16.0
numpy==1.26.0
pyserial==3.5
pyzbar==0.1.9
psycopg2==2.9.10